examples/requirements.txt
# Requirements for Cognitive Memory Layer Examples
# Install with: pip install -r examples/requirements.txt

# py-cml: cognitive-memory-layer (install from packages/py-cml in monorepo)
cognitive-memory-layer>=0.1.0

# Load repo root .env
python-dotenv>=1.0

# HTTP client (standalone_demo uses raw httpx)
httpx>=0.28.0

# OpenAI integration
openai>=1.0

# Anthropic Claude integration
anthropic>=0.18.0

# LangChain integration
langchain>=0.3.0,<0.4.0
langchain-core>=0.3.0,<0.4.0
langchain-openai>=0.3.0,<0.4.0

# Streamlit chatbot (ollama_chatbot_app.py)
streamlit>=1.28.0
